Feat(API Development): #NGGI-43 Update package name
