+++ green-global-initiative-api/src/main/java/com/atu/green_global_initiative_api/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.atu.devops.exception;
+package com.atu.green_global_initiative_api.exception;
 
 public class ResourceNotFoundException  extends RuntimeException {
