+++ green-global-initiative-api/src/main/java/com/atu/green_global_initiative_api/model/dao/UserDetails.java
@@ -1,24 +1,48 @@
-package com.atu.devops.dto;
+package com.atu.green_global_initiative_api.model.dao;
 
-import com.atu.devops.model.dao.ApplicationDetails;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.*;
 import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
 import java.util.List;
 
+@Entity
 @Data
 @Getter
 @Setter
-public class UserDetailsDto {
+@Table(name="user_details")
+public class UserDetails {
+    @Id
+    @JsonProperty("user_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="user_id", unique = true, nullable = false)
     private int userId;
+    @JsonProperty("first_name")
+    @Column(name="first_name",nullable = false)
     private String firstName;
+    @JsonProperty("last_name")
+    @Column(name="last_name",nullable = false)
     private String lastName;
+    @JsonProperty("email")
+    @Column(name="email",nullable = false)
     private String email;
+    @JsonProperty("password")
+    @Column(name="password",nullable = false)
+    private String password;
+    @JsonProperty("isAdmin")
+    @Column(name="isAdmin",nullable = false)
     private boolean isAdmin;
+    @JsonProperty("last_login")
+    @Column(name="last_login",nullable = false)
     private String lastLogin;
 
-    List<ApplicationDetails> applicationDetails;
+//    @OneToMany(fetch = FetchType.LAZY, mappedBy = "userDetails")
+//    private List<ApplicationDetails> applicationDetailsList;
+
+    @OneToMany(mappedBy = "userDetails", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
+    private List<ApplicationDetails> applicationDetails;
 
     public int getUserId() {
         return userId;
@@ -50,6 +74,14 @@ public class UserDetailsDto {
 
     public void setEmail(String email) {
         this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
     }
 
     public boolean isAdmin() {
