+++ green-global-initiative-api/src/main/java/com/atu/green_global_initiative_api/repository/ApplicationDetailsRepo.java
@@ -1,6 +1,6 @@
-package com.atu.devops.repository;
+package com.atu.green_global_initiative_api.repository;
 
-import com.atu.devops.model.dao.ApplicationDetails;
+import com.atu.green_global_initiative_api.model.dao.ApplicationDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
