+++ green-global-initiative-api/src/main/java/com/atu/green_global_initiative_api/service/GrantsApplicationService.java
@@ -1,9 +1,9 @@
-package com.atu.devops.service;
+package com.atu.green_global_initiative_api.service;
 
 
-import com.atu.devops.dto.ApplicationDetailsDto;
-import com.atu.devops.model.dao.ApplicationDetails;
-import com.atu.devops.model.dao.request.ApplicationUpdateRequest;
+import com.atu.green_global_initiative_api.dto.ApplicationDetailsDto;
+import com.atu.green_global_initiative_api.model.dao.ApplicationDetails;
+import com.atu.green_global_initiative_api.model.dao.request.ApplicationUpdateRequest;
 
 import java.util.List;
