+++ green-global-initiative-api/src/test/java/com/atu/devops/NausicaaGreenInitiativeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.atu.devops;
+package com.atu.green_global_initiative_api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
